+++ src/engine/helpers/customEases.js
@@ -1,0 +1,9 @@
+import CustomEase from 'vendors/CustomEase';
+
+export const EASE_BE_HAPPY = 'EASE_BE_HAPPY';
+export const EASE_BE_SAD = 'EASE_BE_SAD';
+
+CustomEase.create(
+  EASE_BE_HAPPY,
+  'M0,0.004 C0,0.004 0.055,0.444 0.175,0.444 0.294,0.444 0.332,0 0.332,0 0.332,0 0.323,0.97 0.618,1 0.9,1.028 1,0 1,0',
+);